requirements.txt
django
django-filter
djangorestframework
django-cors-headers
